Move LevelStatus to utils, as it is more Util then necessary class object.
