+++ src/main/java/world/bentobox/challenges/utils/LevelStatus.java
@@ -1,4 +1,4 @@
-package world.bentobox.challenges;
+package world.bentobox.challenges.utils;
 
 import world.bentobox.challenges.database.object.ChallengeLevel;
